feat: Add scientific committee view
Introduce a new view for the scientific committee, including member details and expertise. Update `ViewState` type and `LandingPage` to accommodate navigation to this new view. The `App` component is updated to conditionally render the `ScientificCommittee` component when the user is not logged in and the view is set to 'committee'.